+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "build lowest version" to the CI workflow, so now the tests are run against both the lowest and the highest dependencies of composer packages
 - Added a new "static" workflow to add phpstan and psalm to perform some static-analysis on the codebase
 - Added a "checks" workflow to perform some additional checks on the codebase
+- Support client, with list issues, get issue, create issue
 
 ## [0.2.4] - 2021-02-11
 ### Added